+++ PhotoGallery.Domain/Entities/Hashtag.cs
@@ -1,0 +1,8 @@
+﻿namespace PhotoGallery.Domain.Entities
+{
+    public class Hashtag
+    {
+        public int Id { get; set; }
+        public string Value { get; set; } = string.Empty;
+    }
+}